big fix

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: jusilanc <jusilanc@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/03 12:14:43 by jusilanc          #+#    #+#             */
-/*   Updated: 2023/04/03 12:18:25 by jusilanc         ###   ########.fr       */
+/*   Updated: 2023/04/05 19:01:00 by jusilanc         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str && *str++)
